requirements.txt
# Core dependencies
flask==3.0.0
kafka-python==2.0.2
apache-flink==1.18.0
redis==5.0.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0